Fix IDT and GDT tests
The problem with the tests was that the tests provided their own
_idt_set and _gdt_set functions to sort of "mock" to calls to LIDT
and LGDT.  Before this patch, both functions had static linkage,
so calls to each in the same module would always resolve to that
static function.  So I removed the static linkage in the interest
of having a working test suite.

The other fix was that the _idt_set and _gdt_set functions in the
actual source would interfere with the mock versions, and give
the linker a stomachache; the solution was to tell the linker
to allow multiple definitions.